+++ pure_python/vector.py
